+++ src/app/components/ModalMeta.tsx
@@ -6,7 +6,7 @@ import { TextInput } from "react-native-gesture-handler";
 import ErrorMessage from "./ErrorMessage";
 import { validateValue } from "../shared/helpers/modal.helper";
 import ModalButtons from "./ModalButtons";
-
+import styles from "./style/stylesModal";
 interface IProps {
   visible: boolean;
   callbackFn: (valor: string) => unknown;
@@ -72,50 +72,3 @@ export default function ModalMeta({
     </Modal>
   );
 }
-
-const styles = StyleSheet.create({
-  centeredView: {
-    flexDirection: "column",
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#00000098",
-  },
-  modal: {
-    flexDirection: "row",
-    marginBottom: 30,
-  },
-  erroValor: {
-    padding: 5,
-  },
-  input: {
-    flexDirection: "column",
-    alignItems: "center",
-  },
-  textInput: {
-    fontSize: 40,
-    width: 150,
-    marginLeft: 15,
-    textAlign: "center",
-  },
-  modalView: {
-    margin: 20,
-    backgroundColor: "white",
-    borderRadius: 20,
-    padding: 35,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 5,
-  },
-  modalText: {
-    marginBottom: 35,
-    textAlign: "center",
-    fontWeight: "bold",
-  },
-});
